backend/pyproject.toml
[project]
name = "ai-coo-shopify-agent-backend"
version = "0.1.0"
description = "Backend for AI COO Shopify Agent - Shopify merchant analytics"
readme = "README.md"
requires-python = ">=3.9"

# Dependencies listed here for reference, but we'll use pip/uv install directly
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.23.2",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.350",
    "langchain-openai>=0.0.2",
    "langchain-anthropic>=0.1.1",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.20",
    "langgraph>=0.0.15",
    "pandas>=2.1.0",
    "faker>=19.3.1",
    "pydantic>=2.3.0",
    "ragas>=0.0.20",
    "python-multipart>=0.0.6",
]

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build" 